Enable and configure wpa_supplicant with dhcpcd
The installer was originally written when the dhcpcd wpa_supplicant hook
was enabled. In early 2019, the hook was disabled and it appears that
the installer has been unable to configure WiFi since then.

To address this, enable the wpa_supplicant service, configure it, and
reload the service when dhcpcd is used for network connectivity.

Closes #5